rag-api/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-community
langchain-google-genai
pymongo
python-dotenv